moved data loader into adapter utils
